Revamp the whole thing
Instead of having different formats for past, current, and future
events, we now only have a single format. It's all in the same directory
as well, so no need to move things around after each event.

It still needs some work on the speakers data part, but other than that
it's good to go.

Now we only need to start making these happen again.